requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0
pydantic==2.5.0
python-dotenv==1.0.0
langchain==0.1.0
langchain-groq==0.1.0
langgraph==0.0.20
slack-sdk==3.26.1
slack-bolt==1.18.1
requests==2.31.0
python-multipart==0.0.6
celery==5.3.4
redis==5.0.1
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
schedule==1.2.0
jinja2==3.1.2
